Prevent clean files from showing up as untracked
Looking at the code in porcelain.path_to_tree_path,
this looks like there was an attempt at this in
85b5383fb816fece4a36faa538f6477703595139, however
relpath is non-destructive, and has no effect unless
the result is assigned to something. If this were
assigned when path is relative, then relpath would
assume path was at root, and so we must check that
the path is absolute before calling relpath.

Fixes #598